Cleanup some includes and move utility functions to the correct header.
