+++ tests/csvm.cpp
@@ -20,7 +20,7 @@
 #include "custom_test_macros.hpp"  // EXPECT_THROW_WHAT, EXPECT_FLOATING_POINT_EQ, EXPECT_FLOATING_POINT_VECTOR_EQ, EXPECT_FLOATING_POINT_2D_VECTOR_EQ
 #include "naming.hpp"              // naming::real_type_label_type_combination_to_name
 #include "types_to_test.hpp"       // util::{real_type_label_type_combination_gtest, real_type_label_type_combination_gtest}
-#include "utility.hpp"             // util::{redirect_output, temporary_file, instantiate_template_file}
+#include "utility.hpp"             // util::{redirect_output, temporary_file, instantiate_template_file, get_distinct_label}
 
 #include "gtest/gtest.h"  // TEST, EXPECT_EQ, EXPECT_TRUE, EXPECT_FALSE, EXPECT_CALL, ::testing::{Test, An}
