+++ tests/detail/io/libsvm_model_parsing.cpp
@@ -18,8 +18,8 @@
 
 #include "../../custom_test_macros.hpp"  // EXPECT_FLOATING_POINT_EQ, EXPECT_THROW_WHAT
 #include "../../naming.hpp"              // naming::real_type_label_type_combination_to_name
-#include "../../types_to_test.hpp"       // util::{instantiate_template_file, real_type_label_type_combination_gtest}
-#include "../../utility.hpp"             // util::temporary_file, util::redirect_output
+#include "../../types_to_test.hpp"       // util::real_type_label_type_combination_gtest
+#include "../../utility.hpp"             // util::{temporary_file, redirect_output, instantiate_template_file, get_distinct_label}
 
 #include "fmt/core.h"              // fmt::format
 #include "gmock/gmock-matchers.h"  // ::testing::HasSubstr
