+++ tests/naming.hpp
@@ -15,84 +15,47 @@
 
 #include "plssvm/detail/arithmetic_type_name.hpp"  // plssvm::detail::arithmetic_type_name
 #include "plssvm/detail/string_utility.hpp"        // plssvm::detail::replace_all
-#include "plssvm/detail/type_traits.hpp"           // plssvm::detail::always_false_v
+#include "plssvm/detail/type_traits.hpp"           // plssvm::detail::{always_false_v, is_map_v, is_unordered_map_v, is_set_v, is_unordered_set_v, is_vector_v}
 
 #include "exceptions/utility.hpp"  // util::exception_type_name
-#include "types_to_test.hpp"       // util::real_type_label_type_combination
 
 #include "fmt/core.h"     // fmt::format
 #include "fmt/ostream.h"  // directly output types with an operator<< overload using fmt
 #include "gtest/gtest.h"  // ::testing::TestParamInfo
 
-#include <filesystem>     // std::filesystem::path
-#include <map>            // std::map
-#include <set>            // std::set
-#include <string>         // std::string
-#include <string_view>    // std::string_view
-#include <tuple>          // std::get
-#include <type_traits>    // std::is_same_v, std::true_type, std::false_type
-#include <unordered_map>  // std::unordered_map
-#include <unordered_set>  // std::unordered_set
-#include <vector>         // std::vector
+#include <filesystem>   // std::filesystem::path
+#include <string>       // std::string
+#include <string_view>  // std::string_view
+#include <tuple>        // std::tuple, std::get
+#include <type_traits>  // std::is_same_v, std::true_type, std::false_type
 
 namespace naming {
 
 // different type traits used in the naming functions
 namespace detail {
 
-// type_trait to check whether the given type is a std::map
-template <typename T>
-struct is_map : std::false_type {};
-template <typename Key, typename Value>
-struct is_map<std::map<Key, Value>> : std::true_type {};
-template <typename T>
-constexpr bool is_map_v = is_map<T>::value;
-
-// type_trait to check whether the given type is a std::unordered_map
-template <typename T>
-struct is_unordered_map : std::false_type {};
-template <typename Key, typename Value>
-struct is_unordered_map<std::unordered_map<Key, Value>> : std::true_type {};
-template <typename T>
-constexpr bool is_unordered_map_v = is_unordered_map<T>::value;
-
-// type_trait to check whether the given type is a std::set
-template <typename T>
-struct is_set : std::false_type {};
-template <typename Key>
-struct is_set<std::set<Key>> : std::true_type {};
-template <typename T>
-constexpr bool is_set_v = is_set<T>::value;
-
-// type_trait to check whether the given type is a std::unordered_set
-template <typename T>
-struct is_unordered_set : std::false_type {};
-template <typename Key>
-struct is_unordered_set<std::unordered_set<Key>> : std::true_type {};
-template <typename T>
-constexpr bool is_unordered_set_v = is_unordered_set<T>::value;
-
-// type_trait to check whether the given type is a std::vector
-template <typename T>
-struct is_vector : std::false_type {};
-template <typename T>
-struct is_vector<std::vector<T>> : std::true_type {};
-template <typename T>
-constexpr bool is_vector_v = is_vector<T>::value;
-
-// type_trait to check whether the given type is a std::tuple
+/**
+ * @brief Type trait to check whether @p T is a `std::tuple`.
+ * @tparam T the type to check
+ */
 template <typename T>
 struct is_tuple : std::false_type {};
+/**
+ * @copybrief naming::detail::is_tuple
+ */
 template <typename... Args>
 struct is_tuple<std::tuple<Args...>> : std::true_type {};
+/**
+ * @copybrief naming::detail::is_tuple
+ */
 template <typename T>
 constexpr bool is_tuple_v = is_tuple<T>::value;
 
 }  // namespace detail
 
-////////////////////////////////////////////////////////////////////////////////
-////                     PRETTY PRINT TYPED_TEST TYPES                      ////
-////////////////////////////////////////////////////////////////////////////////
+//*************************************************************************************************************************************//
+//                                                    PRETTY PRINT TYPED_TEST TYPES                                                    //
+//*************************************************************************************************************************************//
 // detail/utility.cpp
 /**
  * @brief A class used to map a std::map or std::unordered_map to a readable name in the GTest test case name.
@@ -101,9 +64,9 @@ class map_types_to_name {
   public:
     template <typename T>
     static std::string GetName(int) {
-        if constexpr (detail::is_map_v<T>) {
+        if constexpr (plssvm::detail::is_map_v<T>) {
             return "map";
-        } else if constexpr (detail::is_unordered_map_v<T>) {
+        } else if constexpr (plssvm::detail::is_unordered_map_v<T>) {
             return "unordered_map";
         } else {
             static_assert(plssvm::detail::always_false_v<T>, "Invalid type for name mapping provided!");
@@ -117,9 +80,9 @@ class set_types_to_name {
   public:
     template <typename T>
     static std::string GetName(int) {
-        if constexpr (detail::is_set_v<T>) {
+        if constexpr (plssvm::detail::is_set_v<T>) {
             return "set";
-        } else if constexpr (detail::is_unordered_set_v<T>) {
+        } else if constexpr (plssvm::detail::is_unordered_set_v<T>) {
             return "unordered_set";
         } else {
             static_assert(plssvm::detail::always_false_v<T>, "Invalid type for name mapping provided!");
@@ -133,7 +96,7 @@ class vector_types_to_name {
   public:
     template <typename T>
     static std::string GetName(int) {
-        if constexpr (detail::is_vector_v<T>) {
+        if constexpr (plssvm::detail::is_vector_v<T>) {
             return "vector";
         } else {
             static_assert(plssvm::detail::always_false_v<T>, "Invalid type for name mapping provided!");
@@ -185,7 +148,7 @@ class real_type_to_name {
     }
 };
 /**
- * @brief A class used to map all label types to a readable name in the GTest test case name.
+ * @brief A class used to map all legal label types to a readable name in the GTest test case name.
  */
 class label_type_to_name {
   public:
@@ -220,14 +183,14 @@ class real_type_kernel_function_to_name {
     }
 };
 
-////////////////////////////////////////////////////////////////////////////////
-////                   PRETTY PRINT PARAMETERIZED TESTS                     ////
-////////////////////////////////////////////////////////////////////////////////
+//*************************************************************************************************************************************//
+//                                                   PRETTY PRINT PARAMETERIZED TESTS                                                  //
+//*************************************************************************************************************************************//
 namespace detail {
 
 /**
  * @brief Escape some characters of the string such that GTest accepts it as test case name.
- * @details Replaces all "-" with "_M_" (for Minus), all " " with "_W_" (for Whitespace), and "." with "_D_" (for dot).
+ * @details Replaces all "-" with "_M_" (for Minus), all " " with "_W_" (for Whitespace), "." with "_D_" (for dot), and "/" with "_".
  *          If the resulting string would be empty, returns a string containing "EMPTY".
  * @param[in] sv the string to escape for GTest
  * @return the escaped test case name (`[[nodiscard]]`)
@@ -282,7 +245,7 @@ template <typename T>
 // detail/cmd/parameter_*.cpp -> parser_predict, parser_scale, parser_train
 /**
  * @brief Generate a test case name using a command line flag and value combination.
- * @details Replaces all "-" in a flag with "" and escapes the value string using `naming::detail::escape_string()`.
+ * @details Replaces all "-" in a flag with "" and escapes the value string using naming::detail::escape_string().
  * @tparam T the test suite type
  * @param[in] param_info the parameters to aggregate and escape
  * @return the test case name (`[[nodiscard]]`)
